% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadatasample-data.R
\docType{data}
\name{metadatasample}
\alias{metadatasample}
\title{Metadata on samples from OTU table}
\format{A data frame with 461 rows and 3 variables:
\describe{
  \item{sample}{sample ID corresponding to a variable name from \code{otusample}}
  \item{subject}{subject name}
  \item{collection_date}{date the sample was collected, in the format yyyy-mm-dd}
}}
\usage{
data(metadatasample)
}
\description{
Metadata providing subject names and collection dates of samples from the
OTU table.
}
\keyword{datasets}
