% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadatasample_baboon-data.R
\docType{data}
\name{metadatasample_baboon}
\alias{metadatasample_baboon}
\title{Metadata on samples from OTU table of wild baboon data}
\format{
A data frame with 276 rows and 7 variables:
\describe{
  \item{subject}{subject name}
  \item{sample}{sample ID corresponding to a variable name from \code{otusample_baboon}}
  \item{collection_date}{number of days into the study the sample was collected}
  \item{sex}{if subject was male or female}
  \item{season}{if sample was collected in the wet or dry season}
  \item{rain_month_mm}{amount of rainfall in mm for month prior to sample collection}
  \item{diet_PC1}{a measure of dietary composition for a sample}
}
}
\usage{
data(metadatasample_baboon)
}
\description{
Metadata providing subject names and collection dates of samples from the
OTU table.
}
\keyword{datasets}
