% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otusample-data.R
\docType{data}
\name{otusample}
\alias{otusample}
\title{Sample OTU table}
\format{A data frame with 8814 rows and 1781 variables.}
\usage{
data(otusample)
}
\description{
Sample OTU table with 6 subjects. The first variable contains OTU IDs. Each subsequent column represents a
unique sample collected at a specific time from one of six subjects'
microbiomes. Entries represent number of reads per sample per OTU, out of a
total of 30,000 reads for each sample.
}
\keyword{datasets}
