% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otusample_baboon-data.R
\docType{data}
\name{otusample_baboon}
\alias{otusample_baboon}
\title{Sample OTU table}
\format{
A data frame with 2922 rows and 277 variables.
}
\usage{
data(otusample_baboon)
}
\description{
Sample OTU-format table with 6 subjects. The first variable contains OTU IDs.
Each subsequent column represents a unique sample collected at a specific time
from one of 6 subjects' microbiomes. Entries represent number of reads per sample per OTU.
}
\keyword{datasets}
