% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otusample_diet-data.R
\docType{data}
\name{otusample_diet}
\alias{otusample_diet}
\title{Sample OTU table}
\format{
A data frame with 4583 rows and 484 variables.
}
\usage{
data(otusample_diet)
}
\description{
Sample OTU-format table with 34 subjects. Sequences are from metagenomic data, not 16S,
so the first variable contains IDs assigned to each microbial sequence rather than OTUs. 
Each subsequent column represents a unique sample collected at a specific time
from one of 34 subjects' microbiomes. Entries represent number of reads per sample 
per microbial taxon.
}
\keyword{datasets}
