% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomysample-data.R
\docType{data}
\name{taxonomysample}
\alias{taxonomysample}
\title{Taxonomy information for OTUs}
\format{A data frame with 8814 rows and 7 variables.}
\usage{
data(taxonomysample)
}
\description{
Taxonomy information for each OTU listed in the sample OTU table. The first
variable contains OTU IDs, as listed in \code{otusample}, and subsequent
columns provide taxonomic clasification up to Genus, or the most specific
level possible for a given OTU. OTUs that are classified more broadly have
\code{NA} values for narrower taxonomic levels that do not apply.
}
\keyword{datasets}
