% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomysample_baboon-data.R
\docType{data}
\name{taxonomysample_baboon}
\alias{taxonomysample_baboon}
\title{Taxonomy information for OTUs}
\format{
A data frame with 2922 rows and 7 variables.
}
\usage{
data(taxonomysample_baboon)
}
\description{
Taxonomy information for each OTU ID listed in the sample OTU table. The first
variable contains OTU IDs, as listed in \code{otusample_baboon}, and subsequent
columns provide taxonomic clasification up to Genus, or the most specific
level possible for a given taxon. Taxo IDs that are classified more broadly have
\code{NA} values for narrower taxonomic levels that do not apply.
}
\keyword{datasets}
