% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomysample_diet-data.R
\docType{data}
\name{taxonomysample_diet}
\alias{taxonomysample_diet}
\title{Taxonomy information for OTUs}
\format{
A data frame with 4583 rows and 2 variables.
}
\usage{
data(taxonomysample_diet)
}
\description{
Taxonomy information for each taxon ID listed in the sample OTU table. The first
variable contains taxon IDs, as listed in \code{otusample_diet}, and the subsequent
column provides taxonomic clasification up to Genus, or the most specific
level possible for a given taxon. Taxo IDs that are classified more broadly have
\code{NA} values for narrower taxonomic levels that do not apply.
}
\keyword{datasets}
