% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadatasample_diet-data.R
\docType{data}
\name{metadatasample_diet}
\alias{metadatasample_diet}
\title{Metadata on samples from OTU (or other lowest taxonomic level) table of human diet data}
\format{
A data frame with 483 rows and 4 variables:
\describe{
  \item{subject}{subject name}
  \item{sample}{sample ID corresponding to a variable name from \code{otusample_diet}}
  \item{collection_date}{number of days into the study the sample was collected}
  \item{Supplement}{metadata variable indicating if subject was in the EVOO or MCT group}
}
}
\usage{
data(metadatasample_diet)
}
\description{
Metadata providing subject names and collection dates of samples from the
OTU table.
}
\keyword{datasets}
